refactor(overlays): split inline image module
