+++ src/app/overlays/inline_image/window.rs
@@ -1,0 +1,54 @@
+use super::TerminalWindowSize;
+use crossterm::terminal;
+
+pub(super) fn query_terminal_window_size() -> Option<TerminalWindowSize> {
+    let terminal_size = terminal::window_size().ok();
+    let (cells_width, cells_height) = terminal_size
+        .as_ref()
+        .map(|size| (size.columns, size.rows))
+        .or_else(|| terminal::size().ok())?;
+    let (pixels_width, pixels_height) = terminal_size
+        .as_ref()
+        .and_then(|size| {
+            let width = u32::from(size.width);
+            let height = u32::from(size.height);
+            (width > 0 && height > 0).then_some((width, height))
+        })
+        .unwrap_or_else(|| fallback_window_size_pixels(cells_width, cells_height));
+    Some(TerminalWindowSize {
+        cells_width,
+        cells_height,
+        pixels_width,
+        pixels_height,
+    })
+}
+
+#[cfg(test)]
+fn parse_window_size(output: &str) -> Option<(u32, u32)> {
+    let trimmed = output.trim();
+    let (width, height) = trimmed.split_once('x')?;
+    Some((width.parse().ok()?, height.parse().ok()?))
+}
+
+fn fallback_window_size_pixels(cells_width: u16, cells_height: u16) -> (u32, u32) {
+    (
+        u32::from(cells_width.max(1)) * 8,
+        u32::from(cells_height.max(1)) * 16,
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn parse_window_size_reads_pixel_dimensions() {
+        assert_eq!(parse_window_size("1575x919\n"), Some((1575, 919)));
+    }
+
+    #[test]
+    fn fallback_window_size_pixels_uses_reasonable_cell_defaults() {
+        assert_eq!(fallback_window_size_pixels(100, 40), (800, 640));
+        assert_eq!(fallback_window_size_pixels(0, 0), (8, 16));
+    }
+}
